extract shared pickup-refinement gate helper (DRY, #1625)
Replace the duplicated gh issue view fetch + GH_API_ERROR throw +
MISSING_REFINEMENT_ARTIFACT block throw in move-queue-item.mjs with a single
core helper runPickupRefinementGate used by both queue add and queue move.
Honors the PR non-goal of never adding a second copy. Reword the reconcile
inline comment to not overstate reconcile's relationship to the gate.
